Add, and use, a configuration library.
Summary:
I kinda went overboard here, and support both a configuration json
file (with comments) and commandline-flag overrides.

The main thing I wanted to configure was the color use.  This ended up
resulting in quite a bit of code reorganization, but I like the end
result better.

Test Plan:
I ran the following, and they all showed the same graph (which also
looks the same as before this commit):
   ./visualize_jenkins_perf_data.py deploy/build-webapp:14167 -c jenkins_perf_visualizer/config.json
   python3 ./visualize_jenkins_perf_data.py deploy/build-webapp:14167 -c jenkins_perf_visualizer/config.json
   ./visualize_jenkins_perf_data.py /tmp/jenkins-build-perf-analysis/deploy--build-webapp:14167.data -c jenkins_perf_visualizer/config.json
   python3 ./visualize_jenkins_perf_data.py /tmp/jenkins-build-perf-analysis/deploy--build-webapp:14167.data -c jenkins_perf_visualizer/config.json

I then ran:
   ./download_jenkins_perf_data.py deploy/build-webapp -c jenkins_perf_visualizer/config.json
and saw from the jumbled stdout that it was correctly multithreading.

I hit control-C (actually control-\) halfway through, and did
   ls -l /tmp/jenkins-build-perf-analysis/
to ensure the data files were being downloaded with non-zero size.
I then did
   ./download_jenkins_perf_data.py deploy/build-webapp -c jenkins_perf_visualizer/config.json --download-threads=1
which both tests the commandline-flag override, and also the
single-threaded codepath (which is new).  It worked as expected.

At the end of the run I did
   ls /tmp/jenkins-build-perf-analysis/
and saw a bunch of directories like:
   2020-10.74629/
That directory had two files (symlinks) in it, which is expected since
a single deploy usually builds at least twice.  I did
   cat /tmp/jenkins-build-perf-analysis/2020-10.74629/deploy--build-webapp:14162.data
and it gave content, showing the symlink is set up correctly.

Reviewers: benkraft

Reviewed By: benkraft

Differential Revision: https://phabricator.khanacademy.org/D66513

+++ download_jenkins_perf_data.py
@@ -13,19 +13,17 @@ TODO(csilvers): also provide the ability to expire old data, perhaps having a
 max-directory-size.
 """
 import argparse
-import multiprocessing.pool
+import os
+import time
 
+from jenkins_perf_visualizer import configuration
 from jenkins_perf_visualizer import fetch
 from jenkins_perf_visualizer import jenkins
 
 
-# TODO(csilvers): move to a config file
-KEEPER_RECORD_ID = 'mHbUyJXAmnZyqLY3pMUmjQ'
-
-
 def _download_one_build(param):
     # (The weird parameter format is because this is used by Pool().)
-    (job, build_id, output_dir, jenkins_client, force) = param
+    (job, build_id, output_dir, jenkins_client, grouping_param, force) = param
     print("Fetching %s:%s" % (job, build_id))
     try:
         (_, build_params, build_start_time, outfile) = fetch.fetch_build(
@@ -34,24 +32,27 @@ def _download_one_build(param):
         print("ERROR: skipping %s:%s: %s" % (e.job, e.build_id, e))
         return
 
-    # Now create a symlink organized by date and title.
-    # TODO(csilvers): re-enable this via config settings.
-    '''
+    # Now create a symlink farm grouping together jobs with the same
+    # grouping parameter.
+    if not grouping_param:
+        return
+
     yyyy_mm = time.strftime("%Y-%m", time.localtime(build_start_time))
-    title = build_params.get('REVISION_DESCRIPTION', '<unknown build>')
-    category_dir = os.path.join(output_dir, '%s.%s' % (yyyy_mm, title))
+    grouping_id = build_params.get(grouping_param, 'UNKNOWN_GROUP')
+    category_dir = os.path.join(output_dir, '%s.%s' % (yyyy_mm, grouping_id))
     symlink = os.path.join(category_dir, os.path.basename(outfile))
     if force and os.path.exists(symlink):
         os.unlink(symlink)
     if not os.path.exists(symlink):
-        mkdir_p(category_dir)
+        try:
+            os.makedirs(os.path.dirname(symlink))
+        except OSError as e:
+            pass
         os.symlink(os.path.relpath(outfile, os.path.dirname(symlink)),
                    symlink)
-    '''
 
 
-def download_builds(builds, output_dir, jenkins_username, jenkins_password,
-                    force=False):
+def download_builds(config, builds, force=False):
     """Download and save the data-needed-to-render for builds and jobs.
 
     We ask jenkins what builds it knows about for the given jobs,
@@ -60,33 +61,42 @@ def download_builds(builds, output_dir, jenkins_username, jenkins_password,
     graph this build.
 
     Arguments:
+        config: a config.json object
         builds: a list of either builds or jobs, e..g
                   ["deploy/build-webapp:1214", "deploy/webapp-test"]
                 For jobs (where the build-id is omitted), we fetch all
                 builds for the given job that Jenkins knows about.
-        output_dir: the directory to put all the data files
-        jenkins_username: a valid jenkins username
-        jenkins_password: a valid API token.  If None, fetch from keeper.
         force: if False, don't fetch any builds that already have a
                data-file in output_dir.  If True, fetch everything.
     """
-    # TODO(csilvers): use config options to decide what to do, instead.
-    if jenkins_password:
-        jenkins_client = jenkins.get_client_via_password(
-            jenkins_username, jenkins_password)
-    else:
-        jenkins_client = jenkins.get_client_via_keeper(KEEPER_RECORD_ID)
+    if not config.get('datadir'):
+        raise ValueError("No output dir (--datadir) specified")
+
+    jenkins_client = jenkins.get_client(config)
+    download_args = []
     for build in builds:
         if ':' in build:
             (job, build_id) = build.split(':')
-            build_ids = [build_id]
+            download_args.append(
+                (job, build_id, config['datadir'],
+                 jenkins_client, config.get('groupingParameter'), force)
+            )
         else:
             job = build
-            build_ids = jenkins_client.fetch_all_build_ids(job)
+            for build_id in jenkins_client.fetch_all_build_ids(job):
+                download_args.append(
+                    (job, build_id, config['datadir'],
+                     jenkins_client, config.get('groupingParameter'), force)
+                )
 
-        pool = multiprocessing.pool.ThreadPool(7)  # pool size is arbitrary
-        pool.map(_download_one_build,
-                 [(job, b, output_dir, jenkins, force) for b in build_ids])
+    num_threads = config.get('downloadThreads', 7)  # arbitrary number
+    if num_threads <= 1:
+        for args_tuple in download_args:
+            _download_one_build(args_tuple)
+    else:
+        import multiprocessing.pool  # only import if we need it!
+        pool = multiprocessing.pool.ThreadPool(num_threads)
+        pool.map(_download_one_build, download_args)
 
 
 if __name__ == '__main__':
@@ -95,19 +105,18 @@ if __name__ == '__main__':
         'build_or_job', nargs='+',
         help=("Jenkins builds to fetch, e.g. deploy/build-webapp:1543, "
               "OR jenkins jobs to fetch all builds of, e.g. webapp-test."))
-    parser.add_argument('--jenkins-username',
-                        default='jenkins@khanacademy.org')
-    parser.add_argument('--jenkins-pw',
-                        help=('API token that gives access to build data. '
-                              'If not set, fetch the secret from keeper '
-                              '(record %s)' % KEEPER_RECORD_ID))
-    parser.add_argument('-d', '--output-dir',
-                        default='/tmp/jenkins-build-perf-analysis',
-                        help='Directory to write the output data files')
+
+    # Lets you specify a config file to control everything else.
+    configuration.add_config_arg(parser)
+    # Lets you override the values in the config file on a per-run basis.
+    configuration.add_datadir_arg(parser)
+    configuration.add_download_threads_arg(parser)
+
     parser.add_argument('--force', action='store_true',
                         help=('If set, re-fetch data files even if they '
                               'already exist in output-dir.'))
-    args = parser.parse_args()
 
-    download_builds(args.build_or_job, args.output_dir,
-                    args.jenkins_username, args.jenkins_pw, args.force)
+    args = parser.parse_args()
+    config = configuration.load(args)
+
+    download_builds(config, args.build_or_job, args.force)
